feat: change no-console-log to no-console
