+++ examples/dlint/testdata/issue1145_no_trailing_newline.ts
@@ -1,3 +1,3 @@
 /*---
 ---*/
-var base
+var base;
