+++ examples/dlint/testdata/simple.ts
@@ -1,3 +1,2 @@
 function hello(): any {
-    
 }
